Cleanup of the look
